api/requirements.txt
fastapi==0.128.0
mysql-connector-python==9.6.0
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
python-multipart==0.0.22
uvicorn==0.40.0